feat: refactor PlayerActivity to use helper classes for improved code organization and readability
